refact: arrange code layout (no change)
