+++ .github/workflows/02-test.yaml
@@ -11,7 +11,7 @@ jobs:
         os: [ubuntu-latest]
 
     runs-on: ${{ matrix.os }}
-    
+
     steps:
     - uses: actions/checkout@v3
       with:
