+++ uri_test.go
@@ -195,127 +195,6 @@ func Test_ParseThenString(t *testing.T) {
 	})
 }
 
-func Benchmark_Parse(b *testing.B) {
-	tests := []string{
-		"foo://example.com:8042/over/there?name=ferret#nose",
-		"http://httpbin.org/get?utf8=%e2%98%83",
-		"mailto://user@domain.com",
-		"ssh://user@git.openstack.org:29418/openstack/keystone.git",
-		"https://willo.io/#yolo",
-	}
-
-	b.ReportAllocs()
-	b.ResetTimer()
-
-	for i := 0; i < b.N; i++ {
-		_, _ = Parse(tests[i%5])
-	}
-}
-
-func Benchmark_String(b *testing.B) {
-	tests := []*uri{
-		{
-			"foo", "//example.com:8042/over/there", "name=ferret", "nose",
-			&authorityInfo{"//", "", "example.com", "8042", "/over/there"},
-		},
-		{
-			"http", "//httpbin.org/get", "utf8=\xe2\x98\x83", "",
-			&authorityInfo{"//", "", "httpbin.org", "", "/get"},
-		},
-		{
-			"mailto", "user@domain.com", "", "",
-			&authorityInfo{"//", "user", "domain.com", "", ""},
-		},
-		{
-			"ssh", "//user@git.openstack.org:29418/openstack/keystone.git", "", "",
-			&authorityInfo{"//", "user", "git.openstack.org", "29418", "/openstack/keystone.git"},
-		},
-		{
-			"https", "//willo.io/", "", "yolo",
-			&authorityInfo{"//", "", "willo.io", "", "/"},
-		},
-	}
-
-	b.ReportAllocs()
-	b.ResetTimer()
-
-	for i := 0; i < b.N; i++ {
-		_ = tests[i%5].String()
-	}
-}
-
-func Test_Building(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		uri, uriChanged string
-		name            string
-	}{
-		{
-			"mailto://user@domain.com",
-			"http://yolo@newdomain.com:443",
-			"yolo",
-		},
-		{
-			"https://user@domain.com",
-			"http://yolo2@newdomain.com:443",
-			"yolo2",
-		},
-	}
-
-	t.Run("when building from existing URI", func(t *testing.T) {
-		for _, toPin := range tests {
-			test := toPin
-
-			t.Run(fmt.Sprintf("change to %q", test.uriChanged), func(t *testing.T) {
-				t.Parallel()
-
-				auri, err := Parse(test.uri)
-				require.NoErrorf(t, err,
-					"failed to parse uri: %v", err,
-				)
-
-				nuri := auri.Builder().SetUserInfo(test.name).SetHost("newdomain.com").SetScheme("http").SetPort("443")
-				zuri, ok := nuri.(URI)
-				require.True(t, ok)
-				assert.Equal(t, "//"+test.name+"@newdomain.com:443", zuri.Authority().String())
-				assert.Equal(t, "443", nuri.URI().Authority().Port())
-				val := nuri.String()
-
-				assert.Equalf(t, val, test.uriChanged,
-					"val: %#v", val,
-					"test: %#v", test.uriChanged,
-					"values don't match: %v != %v (actual: %#v, expected: %#v)", val, test.uriChanged,
-				)
-				assert.Equal(t, "http", nuri.URI().Scheme())
-
-				_ = nuri.SetPath("/abcd")
-				assert.Equal(t, "/abcd", nuri.URI().Authority().Path())
-
-				_ = nuri.SetQuery("a=b&x=5").SetFragment("chapter")
-				assert.Equal(t, url.Values{"a": []string{"b"}, "x": []string{"5"}}, nuri.URI().Query())
-				assert.Equal(t, "chapter", nuri.URI().Fragment())
-				assert.Equal(t, test.uriChanged+"/abcd?a=b&x=5#chapter", nuri.URI().String())
-				assert.Equal(t, test.uriChanged+"/abcd?a=b&x=5#chapter", nuri.String())
-			})
-		}
-	})
-
-	t.Run("when building from scratch", func(t *testing.T) {
-		u, _ := Parse("http:")
-		b := u.Builder()
-
-		nu := (b.URI()).(*uri)
-		assert.Nil(t, nu.authority)
-
-		require.NotNil(t, u.Authority())
-		assert.Equal(t, "", u.Authority().UserInfo())
-
-		b = b.SetUserInfo("user:pwd").SetHost("newdomain").SetPort("444")
-		assert.Equal(t, "http://user:pwd@newdomain:444", b.String())
-	})
-}
-
 // TestMoreURI borrows from other URI validators to exercise strict RFC3986
 // conformance (taken from .Net, perl, python, ).
 func TestMoreURI(t *testing.T) {
@@ -860,51 +739,4 @@ func Test_Issue3(t *testing.T) {
 		require.Equal(t, "//etc/hosts", auth.Path())
 		require.Empty(t, auth.Host())
 	})
-}
-
-func TestDNSvsHost(t *testing.T) {
-	for _, scheme := range schemesWithDNS() {
-		require.True(t, UsesDNSHostValidation(scheme))
-	}
-
-	require.False(t, UsesDNSHostValidation("phone"))
-}
-
-func schemesWithDNS() []string {
-	return []string{
-		"dns",
-		"dntp",
-		"finger",
-		"ftp",
-		"git",
-		"http",
-		"https",
-		"imap",
-		"irc",
-		"jms",
-		"mailto",
-		"nfs",
-		"nntp",
-		"ntp",
-		"postgres",
-		"redis",
-		"rmi",
-		"rtsp",
-		"rsync",
-		"sftp",
-		"skype",
-		"smtp",
-		"snmp",
-		"soap",
-		"ssh",
-		"steam",
-		"svn",
-		"tcp",
-		"telnet",
-		"udp",
-		"vnc",
-		"wais",
-		"ws",
-		"wss",
-	}
 }
